Improvements to importer processor assignment
There was this overly optimistic calculation of amount of active CPUs
in DynamicProcessorAssigner where processing steps that had much lower
average processing time than the slowest would be counted with a factor
less than one. This would in some cases lead to an over-assignment
of threads which could lead to OS scheduling problems. This could lead
to a slowdown of the import as a whole.

Also stops unnecessary removing threads from "too fast" steps if
there are still threads that are unassigned, this to make things
a fair amount more stable so that thread counts doesn't jump up and
down a whole lot once they have been initially assigned.
Now instead moves thread from one step to another if bottlenecks
are discovered.

+++ community/kernel/src/test/java/org/neo4j/unsafe/impl/batchimport/staging/StageExecutionTest.java
@@ -51,11 +51,11 @@ public class StageExecutionTest
 
         // THEN
         Pair<Step<?>,Float> fastest = ordered.next();
-        assertEquals( 1f / 2f, fastest.other().floatValue(), 0f );
+        assertEquals( 1f / 2f, fastest.other(), 0f );
         Pair<Step<?>,Float> faster = ordered.next();
-        assertEquals( 1f / 3f, faster.other().floatValue(), 0f );
+        assertEquals( 1f / 3f, faster.other(), 0f );
         Pair<Step<?>,Float> fast = ordered.next();
-        assertEquals( 1f, fast.other().floatValue(), 0f );
+        assertEquals( 1f, fast.other(), 0f );
         assertFalse( ordered.hasNext() );
     }
 
@@ -75,13 +75,13 @@ public class StageExecutionTest
 
         // THEN
         Pair<Step<?>,Float> slowest = ordered.next();
-        assertEquals( 3f, slowest.other().floatValue(), 0f );
+        assertEquals( 3f, slowest.other(), 0f );
         Pair<Step<?>,Float> slower = ordered.next();
-        assertEquals( 2f, slower.other().floatValue(), 0f );
+        assertEquals( 2f, slower.other(), 0f );
         Pair<Step<?>,Float> slow = ordered.next();
-        assertEquals( 1f, slow.other().floatValue(), 0f );
+        assertEquals( 1f, slow.other(), 0f );
         Pair<Step<?>,Float> alsoSlow = ordered.next();
-        assertEquals( 1f, alsoSlow.other().floatValue(), 0f );
+        assertEquals( 1f, alsoSlow.other(), 0f );
         assertFalse( ordered.hasNext() );
     }
 }
